feat: merge levels
